AI fix
It turns out that your checkWinner function didn't work for a few reasons. First of all, It ended the game, remember that this function is used to analyse boards that have not been played yet (It is the AI looking at possible moves).

Secondly the method that you used to return doesn't work due to the nature of forEach();

Note: Just incase you want to make this more efficient, you should use a for loop instead of a forEach() function. This is so that you can return true as soon as it finds a 3 in a row (you cannot return in because of the anonymous function) instead of checking for more lines after it has already found one.